modified

+++ rxvt-config/rxvt-config.sh
@@ -14,24 +14,28 @@
 
 function rxvt_config {
 local DIR=$(pwd)
-sudo apt install -y rxvt-unicode-256color xsel
-type rxvt-unicode
-if [[ $? == 0 ]]; then
-	mkdir ~/GitHub
+local STAT=1
+if [[ -e /etc/redhat-release ]]; then
+	sudo yum install -y rxvt-unicode-256color-ml xsel
+	STAT=$?
+fi
+if [[ -e /etc/debian_version ]]; then
+	sudo apt install -y rxvt-unicode-256color xsel
+	STAT=$?
+fi
+if [[ $STAT == 0 ]]; then
+	mkdir -p ~/GitHub
 	cd ~/GitHub
 	git clone https://github.com/kevydotvinu/dotfiles.git
 	echo "Xresources.rxvt copied" && cp dotfiles/.Xresources.rxvt ~/
 	echo "Xresources.rxvt added to .xinitrc" && \
 		echo "xrdb -merge ~/.Xresources.rxvt" >> ~/.xinitrc
-	echo ".bashrc copied" && cp dotfiles/.bashrc ~/
-	echo ".bash_funcs copied" && cp dotfiles/.bash_funcs ~/
-	echo ".bash_alias copied" && cp dotfiles/.bash_alias ~/
 	echo ".fonts copied" && cp -r dotfiles/.fonts/ ~/
 	echo "Xresources updated" && xrdb -merge ~/.Xresources.rxvt
-	sudo update-alternatives --config x-terminal-emulator
 	cd
 fi
-sudo cp $DIR/clipboard /usr/lib/urxvt/perl/
+sudo cp $DIR/clipboard /usr/lib/urxvt/perl/ || \
+	sudo cp $DIR/clipboard /usr/lib64/urxvt/perl/
 }
 
 # Configure rxvt-unicode
